Enable user commands in PerfViewCollect.  Implement LinuxGCStats and LinuxJITStats user commands.
